changed contribution accordain to a components
