+++ .env
@@ -1,1 +1,1 @@
-NEXT_PUBLIC_BASE_URL=https://api.realdevsquad.com
+NEXT_PUBLIC_BASE_URL=https://staging-api.realdevsquad.com
